centralized validate_spec_for_net_devices
