+++ crates/libcontainer/src/container/init_builder.rs
@@ -186,7 +186,11 @@ impl InitContainerBuilder {
             Err(ErrInvalidSpec::UnsupportedVersion)?;
         }
 
-        Validator::validate_spec(spec)?;
+        let syscall = create_syscall();
+        let is_rootless =
+            utils::rootless_required(&*syscall).map_err(LibcontainerError::OtherIO)?;
+
+        Validator::validate_spec(spec, is_rootless)?;
 
         if let Some(process) = spec.process() {
             if let Some(profile) = process.apparmor_profile() {
@@ -204,15 +208,11 @@ impl InitContainerBuilder {
             }
         }
 
-        let syscall = create_syscall();
-
         if let Some(mounts) = spec.mounts() {
             validate_idmapped_mounts(mounts, spec.linux().as_ref(), &*syscall)?;
         }
 
         utils::validate_spec_for_new_user_ns(spec, &*syscall)?;
-        utils::validate_spec_for_net_devices(spec, &*syscall)
-            .map_err(LibcontainerError::NetDevicesError)?;
 
         Ok(())
     }
