+++ crates/libcontainer/src/container/tenant_builder.rs
@@ -357,17 +357,17 @@ impl TenantContainerBuilder {
             Err(ErrInvalidSpec::UnsupportedVersion)?;
         }
 
-        Validator::validate_spec(spec)?;
-
         let syscall = create_syscall();
+        let is_rootless =
+            utils::rootless_required(&*syscall).map_err(LibcontainerError::OtherIO)?;
+
+        Validator::validate_spec(spec, is_rootless)?;
 
         if let Some(mounts) = spec.mounts() {
             validate_idmapped_mounts(mounts, spec.linux().as_ref(), &*syscall)?;
         }
 
         utils::validate_spec_for_new_user_ns(spec, &*syscall)?;
-        utils::validate_spec_for_net_devices(spec, &*syscall)
-            .map_err(LibcontainerError::NetDevicesError)?;
 
         Ok(())
     }
